base-human-mcp-server/requirements.txt
fastmcp>=2.0.0
openai>=1.0.0
pyyaml>=6.0 